moved user to models dir
