+++ tartarus/tartarus/auth.py
@@ -1,7 +1,7 @@
 import functools
 import jwt
 import json
-
+from .models.User import User
 from flask import (
     Blueprint, flash, g, redirect, render_template, request, session, url_for
 )
@@ -10,31 +10,6 @@ from werkzeug.security import check_password_hash, generate_password_hash
 from tartarus.db import get_db
 
 bp = Blueprint('auth', __name__, url_prefix='/auth')
-
-class User:
-    __id = 0
-    __name = ""
-    __email = ""
-    __password = ""
-    __permissions = 0
-    def __init__(self, name, email, password, permissions = 0):
-        self.__name = name
-        self.__email = email
-        self.__password = password
-        self.__permissions = permissions
-    
-    def getId(self):
-        return self.__id
-    def getName(self):
-        return self.__name
-    def getEmail(self):
-        return self.__email
-    def getPermissions(self):
-        return self.__permissions
-    def getPassword(self):
-        return self.__password
-    def setId(self, id):
-        self.__id = id
 
 @bp.route('/newuser', methods=['GET', 'POST'])
 def add_user():
